refactor(qem): extract ZNE into zne.py and derive QEMStage.consumes_dag_bodies
Move ZNE folding/extrapolation out of qem.py into a dedicated zne.py, leaving qem.py as QEMProtocol/_NoMitigation plus new applies_to and requires_bound_params hooks. QEMStage now overrides the existing consumes_dag_bodies property — computed at construction from whether the protocol mitigates — and keys its validation/no-op logic off it instead of inline _NoMitigation checks.
