+++ tests/circuits/test_zne.py
@@ -2,17 +2,35 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 
-"""Tests for GlobalFoldPass, LocalFoldPass, and PauliTwirlPass."""
+"""Tests for divi.circuits.zne."""
 
+import copy
 import random
 
 import pytest
 from qiskit import QuantumCircuit
 from qiskit.circuit import Parameter
+from qiskit.converters import circuit_to_dag, dag_to_circuit
 from qiskit.quantum_info import Operator
 from qiskit.transpiler import PassManager
 
-from divi.circuits._qem_passes import GlobalFoldPass, LocalFoldPass, PauliTwirlPass
+from divi.circuits.qem import QEMProtocol
+from divi.circuits.zne import (
+    ZNE,
+    GlobalFoldPass,
+    LinearExtrapolator,
+    LocalFoldPass,
+    RichardsonExtrapolator,
+)
+from divi.pipeline.abc import ResultFormat
+
+
+@pytest.fixture
+def bell_dag():
+    qc = QuantumCircuit(2)
+    qc.h(0)
+    qc.cx(0, 1)
+    return circuit_to_dag(qc)
 
 
 class TestGlobalFoldPass:
@@ -55,7 +73,7 @@ class TestGlobalFoldPass:
 
         folded = PassManager([GlobalFoldPass(3.0)]).run(qc)
 
-        assert set(p.name for p in folded.parameters) == {"theta", "phi"}
+        assert {p.name for p in folded.parameters} == {"theta", "phi"}
         binding = {theta: 0.4, phi: 0.2}
         assert Operator(folded.assign_parameters(binding)).equiv(
             Operator(qc.assign_parameters(binding))
@@ -68,11 +86,11 @@ class TestGlobalFoldPass:
     @pytest.mark.parametrize(
         "scale,d,expected_size",
         [
-            (1.5, 4, 6),  # k=0, n=1 → 4 + 2*1
-            (2.0, 4, 8),  # k=0, n=2 → 4 + 2*2
-            (2.5, 4, 10),  # k=0, n=3 → 4 + 2*3
-            (3.5, 4, 14),  # k=1, n=1 → 4*3 + 2*1
-            (4.0, 4, 16),  # k=1, n=2 → 4*3 + 2*2
+            (1.5, 4, 6),
+            (2.0, 4, 8),
+            (2.5, 4, 10),
+            (3.5, 4, 14),
+            (4.0, 4, 16),
         ],
     )
     def test_fractional_scale_gate_count(self, scale, d, expected_size):
@@ -96,15 +114,13 @@ class TestGlobalFoldPass:
         assert Operator(folded).equiv(Operator(qc))
 
     def test_fractional_scale_folds_the_tail(self):
-        """scale=1.5 on a 4-gate linear chain folds only the last gate."""
         qc = QuantumCircuit(1)
-        qc.h(0)  # g1
-        qc.rz(0.5, 0)  # g2
-        qc.rx(0.4, 0)  # g3
-        qc.ry(0.3, 0)  # g4 — tail, unambiguous in a linear chain
+        qc.h(0)
+        qc.rz(0.5, 0)
+        qc.rx(0.4, 0)
+        qc.ry(0.3, 0)
         folded = PassManager([GlobalFoldPass(1.5)]).run(qc)
         names = [inst.operation.name for inst in folded.data]
-        # Original 4 gates, then L† = [ry†], then L = [ry].
         assert names == ["h", "rz", "rx", "ry", "ry", "ry"]
         assert Operator(folded).equiv(Operator(qc))
 
@@ -118,7 +134,7 @@ class TestGlobalFoldPass:
 
         folded = PassManager([GlobalFoldPass(2.5)]).run(qc)
 
-        assert set(p.name for p in folded.parameters) == {"theta", "phi"}
+        assert {p.name for p in folded.parameters} == {"theta", "phi"}
         binding = {theta: 0.4, phi: 0.2}
         assert Operator(folded.assign_parameters(binding)).equiv(
             Operator(qc.assign_parameters(binding))
@@ -126,8 +142,7 @@ class TestGlobalFoldPass:
 
 
 class TestLocalFoldPass:
-    """Spec: LocalFoldPass replaces each gate with G·(G†·G)^k, with partial
-    folding for fractional scale factors."""
+    """Spec: LocalFoldPass replaces each gate with G·(G†·G)^k."""
 
     @pytest.fixture
     def four_gate_qc(self):
@@ -159,10 +174,10 @@ class TestLocalFoldPass:
     @pytest.mark.parametrize(
         "scale,d,expected_size",
         [
-            (1.5, 4, 6),  # k=0, n=1 → 4 + 2*1
-            (2.0, 4, 8),  # k=0, n=2 → 4 + 2*2
-            (2.5, 4, 10),  # k=0, n=3 → 4 + 2*3
-            (3.5, 4, 14),  # k=1, n=1 → 4*3 + 2*1
+            (1.5, 4, 6),
+            (2.0, 4, 8),
+            (2.5, 4, 10),
+            (3.5, 4, 14),
         ],
     )
     def test_fractional_scale_gate_count(self, scale, d, expected_size):
@@ -185,14 +200,13 @@ class TestLocalFoldPass:
 
         folded = PassManager([LocalFoldPass(2.0, rng=random.Random(0))]).run(qc)
 
-        assert set(p.name for p in folded.parameters) == {"theta", "phi"}
+        assert {p.name for p in folded.parameters} == {"theta", "phi"}
         binding = {theta: 0.4, phi: 0.2}
         assert Operator(folded.assign_parameters(binding)).equiv(
             Operator(qc.assign_parameters(binding))
         )
 
     def test_non_unitary_ops_are_skipped(self):
-        """Barriers/measurements/resets must not be folded."""
         qc = QuantumCircuit(2, 2)
         qc.h(0)
         qc.cx(0, 1)
@@ -203,12 +217,10 @@ class TestLocalFoldPass:
         op_counts = folded.count_ops()
         assert op_counts.get("barrier", 0) == 1
         assert op_counts.get("measure", 0) == 2
-        # Unitary gates (h, cx) tripled; non-unitary untouched.
         assert op_counts["h"] == 3
         assert op_counts["cx"] == 3
 
     def test_rng_reproducibility(self, four_gate_qc):
-        """Same seed → identical folded structure."""
         a = PassManager([LocalFoldPass(1.5, rng=random.Random(7))]).run(four_gate_qc)
         b = PassManager([LocalFoldPass(1.5, rng=random.Random(7))]).run(four_gate_qc)
         assert [inst.operation.name for inst in a.data] == [
@@ -225,57 +237,46 @@ class TestLocalFoldPass:
 
     @pytest.mark.parametrize("selection", ["from_left", "from_right", "random"])
     def test_selection_preserves_unitary_and_size(self, four_gate_qc, selection):
-        """All selection strategies yield the same gate count and unitary."""
         folded = PassManager(
             [LocalFoldPass(2.0, selection=selection, rng=random.Random(0))]
         ).run(four_gate_qc)
-        assert folded.size() == 8  # d=4, s=2 → 4 + 2*2
+        assert folded.size() == 8
         assert Operator(folded).equiv(Operator(four_gate_qc))
 
     def test_from_left_folds_prefix(self):
-        """from_left: only the first n gates should have folded neighbors."""
         qc = QuantumCircuit(1)
         for _ in range(4):
             qc.rx(0.3, 0)
-        # s=1.5, d=4 → k=0, n=1 → first gate folded (3 gates), rest untouched.
         folded = PassManager([LocalFoldPass(1.5, selection="from_left")]).run(qc)
-        # First 3 instructions = the folded prefix (rx, rx†, rx); next 3 = untouched rx gates.
         names = [inst.operation.name for inst in folded.data]
         assert len(names) == 6
-        assert names[:3] == ["rx", "rx", "rx"]  # G G† G (rx inverse is still "rx")
-        # The remaining 3 are the untouched original rx gates.
+        assert names[:3] == ["rx", "rx", "rx"]
         assert names[3:] == ["rx", "rx", "rx"]
 
     def test_from_right_folds_suffix(self):
-        """from_right: only the last n gates should be folded."""
         qc = QuantumCircuit(2)
         qc.h(0)
         qc.x(1)
         qc.cx(0, 1)
-        qc.rz(0.2, 0)  # this one should be folded
-        # s=1.5, d=4 → n=1, last gate (rz) folded.
+        qc.rz(0.2, 0)
         folded = PassManager([LocalFoldPass(1.5, selection="from_right")]).run(qc)
         names = [inst.operation.name for inst in folded.data]
-        # Expect: h, x, cx, then rz (folded → rz, rz, rz)
         assert names == ["h", "x", "cx", "rz", "rz", "rz"]
 
     def test_exclude_by_name_leaves_gates_untouched(self):
-        """exclude={'cx'} folds everything but cx gates."""
         qc = QuantumCircuit(2)
         qc.h(0)
         qc.cx(0, 1)
         qc.rz(0.3, 1)
         qc.cx(1, 0)
-        # Foldable pool after exclude: [h, rz] → d=2. s=3 → k=1 → each folded to 3 gates.
         folded = PassManager([LocalFoldPass(3.0, exclude={"cx"})]).run(qc)
         counts = folded.count_ops()
-        assert counts["cx"] == 2  # untouched
-        assert counts["h"] == 3  # folded
-        assert counts["rz"] == 3  # folded
+        assert counts["cx"] == 2
+        assert counts["h"] == 3
+        assert counts["rz"] == 3
         assert Operator(folded).equiv(Operator(qc))
 
     def test_exclude_by_arity_shorthand(self):
-        """exclude={'double'} skips all 2-qubit gates."""
         qc = QuantumCircuit(2)
         qc.h(0)
         qc.cx(0, 1)
@@ -290,14 +291,11 @@ class TestLocalFoldPass:
         assert Operator(folded).equiv(Operator(qc))
 
     def test_exclude_affects_effective_scale(self):
-        """With 2 foldable gates out of 4, scale=2 gives n=1 extra fold in the pool."""
         qc = QuantumCircuit(2)
         qc.h(0)
-        qc.cx(0, 1)  # excluded
+        qc.cx(0, 1)
         qc.rz(0.3, 1)
-        qc.cx(1, 0)  # excluded
-        # Foldable d=2, s=2 → k=0, n=round(2*1/2)=1 → 1 of 2 foldable gates gets 1 fold.
-        # Total size: 2 excluded (1 each) + 1 unfolded (1) + 1 folded (3) = 6.
+        qc.cx(1, 0)
         folded = PassManager(
             [LocalFoldPass(2.0, exclude={"cx"}, rng=random.Random(0))]
         ).run(qc)
@@ -305,7 +303,6 @@ class TestLocalFoldPass:
         assert Operator(folded).equiv(Operator(qc))
 
     def test_exclude_unknown_name_is_harmless(self, four_gate_qc):
-        """Unknown op names in exclude are silently no-ops."""
         folded = PassManager([LocalFoldPass(3.0, exclude={"nonexistent_gate"})]).run(
             four_gate_qc
         )
@@ -313,7 +310,6 @@ class TestLocalFoldPass:
         assert Operator(folded).equiv(Operator(four_gate_qc))
 
     def test_exclude_all_gates_returns_untouched(self, four_gate_qc):
-        """Excluding every gate → empty foldable pool → no change."""
         folded = PassManager([LocalFoldPass(3.0, exclude={"single", "double"})]).run(
             four_gate_qc
         )
@@ -321,7 +317,6 @@ class TestLocalFoldPass:
         assert Operator(folded).equiv(Operator(four_gate_qc))
 
     def test_deterministic_selections_ignore_rng(self, four_gate_qc):
-        """from_left / from_right produce identical output regardless of rng."""
         for selection in ("from_left", "from_right"):
             a = PassManager(
                 [LocalFoldPass(1.5, selection=selection, rng=random.Random(0))]
@@ -339,87 +334,185 @@ class TestLocalFoldPass:
         assert folded.size() == 0
 
 
-class TestPauliTwirlPass:
-    """Spec: PauliTwirlPass preserves the ideal unitary (up to global phase)."""
+class TestZNEProtocol:
+    def test_is_qem_protocol(self):
+        assert isinstance(ZNE([1.0, 3.0]), QEMProtocol)
+
+    def test_applies_to_expvals_only(self):
+        zne = ZNE([1.0, 3.0])
+        assert zne.applies_to(ResultFormat.EXPVALS) is True
+        assert zne.applies_to(ResultFormat.PROBS) is False
+        assert zne.applies_to(ResultFormat.COUNTS) is False
+
+    def test_twirl_and_bind_defaults(self):
+        zne = ZNE([1.0, 3.0])
+        assert zne.n_twirls == 0
+        assert zne.requires_bound_params is False
+
+
+class TestZNE:
+    def test_valid_initialization(self):
+        zne = ZNE(scale_factors=[1.0, 3.0, 5.0])
+        assert zne.name == "zne"
+        assert list(zne.scale_factors) == [1.0, 3.0, 5.0]
+        assert isinstance(zne.extrapolator, RichardsonExtrapolator)
+
+    def test_accepts_explicit_extrapolator(self):
+        extrapolator = LinearExtrapolator()
+        zne = ZNE(scale_factors=[1.0, 3.0], extrapolator=extrapolator)
+        assert zne.extrapolator is extrapolator
 
     @pytest.mark.parametrize(
-        "gate_method",
-        ["cx", "cz"],
+        "bad_scale",
+        [
+            "not_a_sequence",
+            [1.0, "foo"],
+        ],
     )
-    def test_twirl_preserves_unitary(self, gate_method):
-        """Twirling CX or CZ gates preserves the overall unitary."""
-        qc = QuantumCircuit(2)
-        qc.h(0)
-        getattr(qc, gate_method)(0, 1)
-        qc.h(1)
-        getattr(qc, gate_method)(1, 0)
+    def test_rejects_invalid_scale_factor_types(self, bad_scale):
+        with pytest.raises(ValueError, match="sequence of real numbers"):
+            ZNE(scale_factors=bad_scale)
 
-        u_orig = Operator(qc)
-        for seed in range(10):
-            rng = random.Random(seed)
-            twirled = PassManager([PauliTwirlPass(rng=rng)]).run(qc)
-            assert Operator(twirled).equiv(
-                u_orig
-            ), f"Seed {seed}: twirled unitary not equivalent"
+    def test_rejects_scale_factor_below_one(self):
+        with pytest.raises(ValueError, match="≥ 1"):
+            ZNE(scale_factors=[0.5, 1.0])
 
-    def test_mixed_cx_cz_preserves_unitary(self):
-        """Circuit mixing CX and CZ gates is correctly twirled."""
-        qc = QuantumCircuit(3)
-        qc.h(0)
-        qc.cx(0, 1)
-        qc.cz(1, 2)
-        qc.cx(2, 0)
+    def test_rejects_non_extrapolator(self):
+        with pytest.raises(ValueError, match="ZNEExtrapolator"):
+            ZNE(scale_factors=[1.0, 3.0], extrapolator="not an extrapolator")
 
-        u_orig = Operator(qc)
-        for seed in range(10):
-            rng = random.Random(seed)
-            twirled = PassManager([PauliTwirlPass(rng=rng)]).run(qc)
-            assert Operator(twirled).equiv(u_orig)
+    def test_expand_returns_one_dag_per_scale(self, bell_dag):
+        zne = ZNE(scale_factors=[1.0, 3.0, 5.0])
+        dags, ctx = zne.expand(bell_dag)
+        assert len(dags) == 3
+        assert ctx["effective_scales"] == (1.0, 3.0, 5.0)
 
-    def test_non_twirl_gates_untouched(self):
-        """Single-qubit gates pass through unmodified."""
-        qc = QuantumCircuit(2)
-        qc.h(0)
-        qc.rz(0.5, 1)
-        qc.ry(0.3, 0)
+    def test_expand_preserves_unitary(self, bell_dag):
+        zne = ZNE(scale_factors=[1.0, 3.0, 5.0])
+        dags, _ = zne.expand(bell_dag)
+        u_orig = Operator(dag_to_circuit(bell_dag))
+        for d in dags:
+            assert Operator(dag_to_circuit(d)).equiv(u_orig)
 
-        rng = random.Random(42)
-        twirled = PassManager([PauliTwirlPass(rng=rng)]).run(qc)
-        assert twirled.size() == qc.size()
-        assert Operator(twirled).equiv(Operator(qc))
+    def test_expand_scales_gate_count(self, bell_dag):
+        zne = ZNE(scale_factors=[1.0, 3.0, 5.0])
+        base = bell_dag.size()
+        dags, _ = zne.expand(bell_dag)
+        assert [d.size() for d in dags] == [base, 3 * base, 5 * base]
 
-    def test_parametric_circuit_preserves_bound_unitary(self):
-        """Twirled parametric circuit produces correct unitary after binding."""
-        theta = Parameter("theta")
-        qc = QuantumCircuit(2)
-        qc.rx(theta, 0)
-        qc.cx(0, 1)
-        qc.rz(theta, 1)
+    def test_reduce_extrapolates_to_zero(self, bell_dag):
+        zne = ZNE(scale_factors=[1.0, 3.0, 5.0], extrapolator=LinearExtrapolator())
+        extrapolated = zne.reduce(
+            [1.0, -1.0, -3.0], {"effective_scales": (1.0, 3.0, 5.0)}
+        )
+        assert extrapolated == pytest.approx([2.0])
 
-        rng = random.Random(0)
-        twirled = PassManager([PauliTwirlPass(rng=rng)]).run(qc)
-        assert set(p.name for p in twirled.parameters) == {"theta"}
+    def test_reduce_falls_back_to_requested_scales_without_context(self, bell_dag):
+        zne = ZNE(scale_factors=[1.0, 3.0, 5.0], extrapolator=LinearExtrapolator())
+        assert zne.reduce([1.0, -1.0, -3.0], {}) == pytest.approx([2.0])
 
-        for val in [0.0, 0.5, 1.2, -0.7]:
-            binding = {theta: val}
-            assert Operator(twirled.assign_parameters(binding)).equiv(
-                Operator(qc.assign_parameters(binding))
-            ), f"theta={val}: bound twirled unitary not equivalent"
+    def test_expand_forwards_effective_scales_to_reduce(self, bell_dag):
+        zne = ZNE(scale_factors=[1.0, 3.0, 5.0], extrapolator=LinearExtrapolator())
+        _, ctx = zne.expand(bell_dag)
+        assert zne.reduce([1.0, -1.0, -3.0], ctx) == pytest.approx([2.0])
 
-    def test_distinct_seeds_produce_distinct_but_equivalent_twirls(self):
-        """Different seeds produce structurally different but unitarily equivalent circuits."""
-        qc = QuantumCircuit(3)
-        qc.cx(0, 1)
-        qc.cx(1, 2)
-        qc.cx(0, 2)
-        qc.cx(1, 0)
+    def test_expand_warns_when_scales_collapse(self, bell_dag):
+        zne = ZNE(scale_factors=[1.5, 2.5, 3.0])
+        with pytest.warns(UserWarning, match="collapse to effective scales"):
+            zne.expand(bell_dag)
 
-        u_orig = Operator(qc)
-        variants = set()
-        for seed in range(20):
-            rng = random.Random(seed)
-            twirled = PassManager([PauliTwirlPass(rng=rng)]).run(qc)
-            variants.add(tuple(inst.operation.name for inst in twirled.data))
-            assert Operator(twirled).equiv(u_orig)
+    def test_reduce_with_per_obs_list(self, bell_dag):
+        zne = ZNE(scale_factors=[1.0, 3.0], extrapolator=LinearExtrapolator())
+        _, ctx = zne.expand(bell_dag, observable=("a", "b"))
+        rows = [[1.0, 2.0], [-1.0, -2.0]]
+        out = zne.reduce(rows, ctx)
+        assert isinstance(out, list)
+        assert out[0] == pytest.approx(2.0)
+        assert out[1] == pytest.approx(4.0)
 
-        assert len(variants) > 1
+
+class TestZNEDryExpand:
+    """ZNE's analytic dry path: same fan-out/context as expand, zero mutation."""
+
+    def test_emits_one_dag_per_scale_without_folding(self, bell_dag):
+        zne = ZNE(scale_factors=[1.0, 3.0, 5.0])
+        base = bell_dag.size()
+        dags, ctx = zne.dry_expand(bell_dag)
+        assert len(dags) == 3
+        assert all(d is bell_dag for d in dags)
+        assert bell_dag.size() == base
+        assert ctx["dag_indices"] == [0, 1, 2]
+
+    def test_effective_scales_match_expand(self, bell_dag):
+        zne = ZNE(scale_factors=[1.0, 3.0, 5.0])
+        _, dry_ctx = zne.dry_expand(copy.deepcopy(bell_dag))
+        _, real_ctx = zne.expand(bell_dag)
+        assert dry_ctx["effective_scales"] == real_ctx["effective_scales"]
+
+    def test_aliased_batch_does_not_compound(self, bell_dag):
+        zne = ZNE(scale_factors=[1.0, 3.0, 5.0])
+        base = bell_dag.size()
+        shared_entries = [bell_dag] * 10
+        for dag in shared_entries:
+            zne.dry_expand(dag)
+        assert bell_dag.size() == base
+
+    def test_warns_when_scales_collapse(self, bell_dag):
+        zne = ZNE(scale_factors=[1.5, 2.5, 3.0])
+        with pytest.warns(UserWarning, match="collapse to effective scales"):
+            zne.dry_expand(bell_dag)
+
+
+class TestLinearExtrapolator:
+    def test_fits_line_through_two_points(self):
+        e = LinearExtrapolator()
+        assert e.extrapolate([1.0, 3.0], [3.0, 7.0]) == pytest.approx(1.0)
+
+    def test_intercept_from_three_points(self):
+        e = LinearExtrapolator()
+        sfs = [1.0, 3.0, 5.0]
+        ys = [4.5, 3.5, 2.5]
+        assert e.extrapolate(sfs, ys) == pytest.approx(5.0)
+
+    def test_rejects_mismatched_lengths(self):
+        with pytest.raises(ValueError, match="lengths disagree"):
+            LinearExtrapolator().extrapolate([1.0, 3.0], [1.0])
+
+    def test_rejects_single_point(self):
+        with pytest.raises(ValueError, match="at least 2"):
+            LinearExtrapolator().extrapolate([1.0], [2.0])
+
+    def test_rejects_nan_input(self):
+        with pytest.raises(ValueError, match="NaN or Inf"):
+            LinearExtrapolator().extrapolate([1.0, 3.0], [float("nan"), 1.0])
+
+    def test_rejects_inf_input(self):
+        with pytest.raises(ValueError, match="NaN or Inf"):
+            LinearExtrapolator().extrapolate([1.0, float("inf")], [1.0, 2.0])
+
+
+class TestRichardsonExtrapolator:
+    def test_interpolates_polynomial_exactly(self):
+        e = RichardsonExtrapolator()
+        sfs = [1.0, 2.0, 3.0]
+        ys = [4.0, 4.0, 2.0]
+        assert e.extrapolate(sfs, ys) == pytest.approx(2.0)
+
+    def test_linear_through_two_points_matches_linear(self):
+        sfs = [1.0, 3.0]
+        ys = [0.5, -0.5]
+        richardson = RichardsonExtrapolator().extrapolate(sfs, ys)
+        linear = LinearExtrapolator().extrapolate(sfs, ys)
+        assert richardson == pytest.approx(linear)
+
+    def test_rejects_mismatched_lengths(self):
+        with pytest.raises(ValueError, match="lengths disagree"):
+            RichardsonExtrapolator().extrapolate([1.0, 3.0], [1.0])
+
+    def test_rejects_duplicate_scale_factors(self):
+        with pytest.raises(ValueError, match="duplicates"):
+            RichardsonExtrapolator().extrapolate([1.0, 3.0, 3.0], [0.5, 0.3, 0.3])
+
+    def test_rejects_nan_input(self):
+        with pytest.raises(ValueError, match="NaN or Inf"):
+            RichardsonExtrapolator().extrapolate([1.0, 3.0], [float("nan"), 1.0])
